# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cautai AI Search Engine API
  description: |
    Cautai is an AI-first search engine designed for both agents and humans. This API provides 
    comprehensive search capabilities, content composition, and citation generation services.
    
    ## Features
    - Advanced search with hybrid ranking (BM25 + semantic similarity)
    - AI-powered content composition and summarization
    - Academic citation generation in multiple formats
    - Multi-language support (English, Romanian)
    - Rate limiting and authentication
    - Real-time search result processing
    
    ## Authentication
    Most endpoints require Bearer token authentication. Obtain tokens through the authentication endpoints.
    
    ## Rate Limiting
    - Search endpoints: 100 requests per minute per IP
    - Composition endpoints: 20 requests per minute per authenticated user
    - Citation endpoints: 50 requests per minute per authenticated user
    
  version: 1.0.0
  contact:
    name: Cautai Support
    url: https://cautai.ro/support
    email: support@cautai.ro
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.cautai.ro/v1
    description: Production server
  - url: https://staging-api.cautai.ro/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
paths:
...
  /compose:
    post:
      summary: Generate AI composition
      description: |
        Generate AI-powered content composition based on search results and user requirements.
        Supports multiple styles and formats for different use cases.
      operationId: compose
      tags:
        - Composition
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompositionRequest'
            examples:
              informative_composition:
                summary: Informative article composition
                value:
                  query: "explain quantum computing"
                  style: "informative"
                  maxLength: 1000
                  language: "en"
              summary_composition:
                summary: Brief summary composition
                value:
                  query: "latest AI trends 2025"
                  style: "summary"
                  maxLength: 300
                  language: "en"
      responses:
        '200':
          description: Composition generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompositionResponse'
        '400':
          description: Invalid composition parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CompositionRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 500
          description: Topic or question for composition
        style:
          type: string
          enum: ["informative", "summary", "detailed", "brief"]
          default: "informative"
          description: Writing style for composition
        maxLength:
          type: integer
          minimum: 100
          maximum: 5000
          default: 1000
          description: Maximum length in characters
        language:
          type: string
          enum: ["en", "ro"]
          default: "en"
          description: Language for composition
        sources:
          type: array
          items:
            type: string
            format: uri
          description: Specific sources to use for composition
    CompositionResponse:
      type: object
      properties:
        compositionId:
          type: string
          format: uuid
          description: Unique identifier for this composition
        query:
          type: string
          description: Original composition query
        composition:
          type: string
          description: Generated content
        sources:
          type: array
          items:
            $ref: '#/components/schemas/CompositionSource'
        processingTime:
          type: integer
          description: Processing time in milliseconds
        metadata:
          type: object
          properties:
            style:
              type: string
            length:
              type: integer
            language:
              type: string
    CompositionSource:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
          format: uri
        relevance:
          type: number
          minimum: 0
          maximum: 1
        excerpt:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            type: object
          description: Detailed error information
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          description: Unique request identifier for debugging
...